requirements.txt
# --- FastAPI and Web Server ---
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.3
aiofiles==23.2.1
python-multipart==0.0.9

# --- LangChain & RAG Components ---
langchain==0.2.1
langchain-community==0.2.1
langchainhub==0.1.15

# --- Vector Store (FAISS) ---
faiss-cpu==1.8.0
sentence-transformers==2.6.1

# --- Hugging Face Models ---
transformers==4.41.2
torch==2.3.0
huggingface-hub==0.23.0

# --- Groq API Integration ---
groq==0.28.0

# --- Utilities & Parsing ---
requests==2.32.3
pandas==2.2.2
feedparser==6.0.11
beautifulsoup4==4.12.3

# --- Email Functionality ---
email-validator==2.1.0.post1
python-dotenv==1.0.1
newsapi-python==0.2.7
beautifulsoup4==4.12.3
requests==2.32.3
